Move the var-builder in a central place.
