+++ candle-examples/examples/bert/main.rs
@@ -4,72 +4,14 @@
 extern crate intel_mkl_src;
 
 use anyhow::{anyhow, Error as E, Result};
-use candle::{safetensors::SafeTensors, DType, Device, Shape, Tensor};
+use candle::{DType, Device, Tensor};
 use candle_hub::{api::sync::Api, Cache, Repo, RepoType};
-use candle_nn::{Embedding, LayerNorm, Linear};
+use candle_nn::{Embedding, LayerNorm, Linear, VarBuilder};
 use clap::Parser;
 use serde::Deserialize;
-use std::collections::HashMap;
 use tokenizers::{PaddingParams, Tokenizer};
 
 const DTYPE: DType = DType::F32;
-
-struct VarBuilder<'a> {
-    safetensors: Option<(HashMap<String, usize>, Vec<SafeTensors<'a>>)>,
-    dtype: DType,
-    device: Device,
-}
-
-impl<'a> VarBuilder<'a> {
-    pub fn from_safetensors(
-        safetensors: Vec<SafeTensors<'a>>,
-        dtype: DType,
-        device: Device,
-    ) -> Self {
-        let mut routing = HashMap::new();
-        for (index, sf) in safetensors.iter().enumerate() {
-            for k in sf.names() {
-                routing.insert(k.to_string(), index);
-            }
-        }
-        Self {
-            safetensors: Some((routing, safetensors)),
-            device,
-            dtype,
-        }
-    }
-
-    pub fn zeros(dtype: DType, device: Device) -> Self {
-        Self {
-            safetensors: None,
-            device,
-            dtype,
-        }
-    }
-
-    pub fn get<S: Into<Shape>>(&self, s: S, tensor_name: &str) -> candle::Result<Tensor> {
-        let s: Shape = s.into();
-        match &self.safetensors {
-            None => Tensor::zeros(s, self.dtype, &self.device),
-            Some((routing, safetensors)) => {
-                // Unwrap or 0  just to let the proper error flow.
-                let index = routing.get(tensor_name).unwrap_or(&0);
-                let tensor = safetensors[*index]
-                    .tensor(tensor_name, &self.device)?
-                    .to_dtype(self.dtype)?;
-                if *tensor.shape() != s {
-                    let msg = format!("shape mismatch for {tensor_name}");
-                    Err(candle::Error::UnexpectedShape {
-                        msg,
-                        expected: s,
-                        got: tensor.shape().clone(),
-                    })?
-                }
-                Ok(tensor)
-            }
-        }
-    }
-}
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Deserialize)]
 #[serde(rename_all = "lowercase")]
@@ -633,7 +575,7 @@ impl Args {
 
         let weights = unsafe { candle::safetensors::MmapedFile::new(weights_filename)? };
         let weights = weights.deserialize()?;
-        let vb = VarBuilder::from_safetensors(vec![weights], DTYPE, device);
+        let vb = VarBuilder::from_safetensors(vec![weights], DTYPE, &device);
         let model = BertModel::load(&vb, &config)?;
         Ok((model, tokenizer))
     }
