+++ candle-examples/examples/whisper/main.rs
@@ -12,13 +12,14 @@ extern crate intel_mkl_src;
 use anyhow::{Error as E, Result};
 use candle::{DType, Device, Tensor};
 use candle_hub::{api::sync::Api, Repo, RepoType};
+use candle_nn::VarBuilder;
 use clap::Parser;
 use rand::{distributions::Distribution, SeedableRng};
 use tokenizers::Tokenizer;
 
 mod audio;
 mod model;
-use model::{Config, VarBuilder, Whisper};
+use model::{Config, Whisper};
 
 const DTYPE: DType = DType::F32;
