+++ candle-examples/examples/whisper/model.rs
@@ -1,67 +1,9 @@
 // We use anyhow rather than candle errors as it provides better support for getting the backtrace
 // back when using RUST_LIB_BACKTRACE=1.
 use anyhow::Result;
-use candle::{safetensors::SafeTensors, DType, Device, Shape, Tensor};
-use candle_nn::{Conv1d, Conv1dConfig, Embedding, LayerNorm, Linear};
+use candle::{Device, Tensor};
+use candle_nn::{Conv1d, Conv1dConfig, Embedding, LayerNorm, Linear, VarBuilder};
 use serde::Deserialize;
-use std::collections::HashMap;
-
-pub struct VarBuilder<'a> {
-    safetensors: Option<(HashMap<String, usize>, Vec<SafeTensors<'a>>)>,
-    dtype: DType,
-    device: Device,
-}
-
-impl<'a> VarBuilder<'a> {
-    pub fn from_safetensors(
-        safetensors: Vec<SafeTensors<'a>>,
-        dtype: DType,
-        device: &Device,
-    ) -> Self {
-        let mut routing = HashMap::new();
-        for (index, sf) in safetensors.iter().enumerate() {
-            for k in sf.names() {
-                routing.insert(k.to_string(), index);
-            }
-        }
-        Self {
-            safetensors: Some((routing, safetensors)),
-            device: device.clone(),
-            dtype,
-        }
-    }
-
-    pub fn zeros(dtype: DType, device: Device) -> Self {
-        Self {
-            safetensors: None,
-            device,
-            dtype,
-        }
-    }
-
-    pub fn get<S: Into<Shape>>(&self, s: S, tensor_name: &str) -> candle::Result<Tensor> {
-        let s: Shape = s.into();
-        match &self.safetensors {
-            None => Tensor::zeros(s, self.dtype, &self.device),
-            Some((routing, safetensors)) => {
-                // Unwrap or 0  just to let the proper error flow.
-                let index = routing.get(tensor_name).unwrap_or(&0);
-                let tensor = safetensors[*index]
-                    .tensor(tensor_name, &self.device)?
-                    .to_dtype(self.dtype)?;
-                if *tensor.shape() != s {
-                    let msg = format!("shape mismatch for {tensor_name}");
-                    Err(candle::Error::UnexpectedShape {
-                        msg,
-                        expected: s,
-                        got: tensor.shape().clone(),
-                    })?
-                }
-                Ok(tensor)
-            }
-        }
-    }
-}
 
 // The names in comments correspond to the original implementation:
 // https://github.com/openai/whisper/blob/f572f2161ba831bae131364c3bffdead7af6d210/whisper/model.py#L17
